pyproject.toml
[tool.poetry]
name = "discordpy-bot-template-with-commands-extension"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
asyncpraw = "^7.4.0"
py-cord = "^1.7.3"
joke-generator = "^1.0.1"
discord = "^1.7.3"
pyjokes = "^0.6.0"
psutil = "^5.8.0"
pypokedex = "^1.6.0"
PyGithub = "^1.55"
animec = "^0.3.3"
DiscordUtils = {extras = ["voice"], version = "^1.3.4"}
python-weather = "^0.3.6"
discord-components = "^2.1.0"
discord-together = "^1.1.1"
IMDbPY = "^2021.4.18"
SQLAlchemy = "1.4.22"
imdb-cli-tool = "^1.0.1"
replit = "^3.2.4"
discordSuperUtils = "^0.2.0"
aiosqlite = "^0.17.0"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
